requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
cryptography==2.3.1
Django==2.1.2
django-rest-knox==3.4.0
djangorestframework==3.9.0
drf-nested-routers==0.91
Faker==0.9.2
idna==2.7
Pillow==5.3.0
pycparser==2.19
pyOpenSSL==18.0.0
python-dateutil==2.7.5
pytz==2018.7
six==1.11.0
text-unidecode==1.2
